[#3] Feat: 상세화면 뷰 구현 및 상세화면 뷰컨트롤러에서 분리
